Use _WIN32 instead of _MSC_VER where appropriate.
_WIN32 is always defined when targeting Windows, whereas _MSC_VER is
only defined for the Visual Studio compiler. As such, _MSC_VER should
only be used for things that are Visual Studio specific, and _WIN32
should be used for things that are Windows specific regardless of
compiler.

This helps to compile Effekseer with MinGW instead of a native Visual
Studio toolchain.